remove unused queue
There was a queue with the purpose to send a new path for newly acquired backup from backup requester to node manager, but it created a strange loop since there is already a queue from node manager to backup requester. Instead, the node manager checks the files in dir to acquire knowledge ob available backups.